pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==1.11.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.11.0",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "colorlog==6.10.1",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "dill==0.4.0",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docling==2.62.0",
    "docling-core==2.51.1",
    "docling-ibm-models==3.10.2",
    "docling-parse==4.7.1",
    "email-validator==2.3.0",
    "et-xmlfile==2.0.0",
    "executing==2.2.1",
    "faker==38.2.0",
    "fastapi==0.121.2",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "greenlet==3.2.4",
    "groq==0.34.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "ipykernel==7.1.0",
    "ipython==9.7.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jsonlines==4.0.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonref==1.1.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.9.1",
    "langchain==1.0.7",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.0.5",
    "langchain-groq==1.0.1",
    "langchain-openai==1.0.3",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.3",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.4",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.43",
    "latex2mathml==3.78.1",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "marko==2.2.1",
    "markupsafe==3.0.3",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mpire==2.10.2",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "multiprocess==0.70.18",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "numpy==2.2.6",
    "omegaconf==2.3.0",
    "openai==2.8.1",
    "opencv-python==4.12.0.88",
    "openpyxl==3.1.5",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "passlib[bcrypt]>=1.7.4",
    "pillow==11.3.0",
    "platformdirs==4.5.0",
    "pluggy==1.6.0",
    "polyfactory==3.0.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "psutil==7.1.3",
    "psycopg2-binary>=2.9.11",
    "pure-eval==0.2.3",
    "pyclipper==1.3.0.post6",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pylatexenc==2.10",
    "pypdf2==3.0.1",
    "pypdfium2==4.30.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart==0.0.20",
    "python-pptx==1.0.2",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "rapidocr==3.4.2",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rpds-py==0.29.0",
    "rtree==1.4.1",
    "safetensors==0.7.0",
    "scipy==1.16.3",
    "semchunk==2.2.2",
    "shapely==2.1.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.44",
    "stack-data==0.6.3",
    "starlette==0.49.3",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.57.1",
    "tree-sitter==0.25.2",
    "tree-sitter-c==0.24.1",
    "tree-sitter-java==0.23.5",
    "tree-sitter-javascript==0.25.0",
    "tree-sitter-python==0.25.0",
    "tree-sitter-typescript==0.23.2",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "wcwidth==0.2.14",
    "xlsxwriter==3.2.9",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
